refactor(db/types): rework database traits to use `gen_proxy` macro, reorganize
