+++ crates/db/types/Cargo.toml
@@ -23,17 +23,26 @@ strata-state.workspace = true
 strata-storage-common.workspace = true
 zkaleido.workspace = true
 
+# Optional: only pulled in by the `proxies` feature, which generates the
+# tokio-backed `*Proxy` database wrappers via `gen_proxy`.
+strata-db-macros = { workspace = true, optional = true }
+
 anyhow.workspace = true
 arbitrary.workspace = true
 bitcoin.workspace = true
 borsh.workspace = true
 serde.workspace = true
 thiserror.workspace = true
+tokio = { workspace = true, optional = true, features = ["rt"] }
+tracing = { workspace = true, optional = true }
 typed-sled.workspace = true
 
 [features]
 default = []
 mocks = []
+# Generates `*Proxy` async database wrappers via `gen_proxy`. Pulls in tokio
+# (for the spawn_blocking dispatch) and tracing (for the instrumented shims).
+proxies = ["dep:strata-db-macros", "dep:tokio", "dep:tracing"]
 
 [dev-dependencies]
 proptest.workspace = true
